test(core): move parser tests to separate file
